Added a random noise grain effect
Added a randomized noise/film grain/static effect to the camera. Tried adding chromatic aberration too, but it didn't work out (has some stuff for that in the code, though).
Effect currently only applied to the character camera, attached to the sprite 2D I added on the camera in the player scene.