Added Airlock door
Airlock now has a door, on the pod side.
The door will close when you first enter the scene. The door will reopen when you walk up to it. It will wait until it's fully closed before opening again. Once the buttons are pressed, the door will remain closed and will not reopen.