Helianth stays moving
The Helianth would "pause" when a player enters a room that isn't the bottom hall, top hall, and stairwell.

I've improved some things and it'll continue to move while the player is in a room

Also it doesn't phase through the staircase door if security is still enabled.

Still really easy to walk into the Helianth when going through doors